feat(code): default seam services
